Abstract mod stuff to mods.js
